fix(tiendas): pulir hallazgos Minor de la revision final del batch P0 #8
- radio_permitido: integer (no numeric) en validacion back/front, para
  que coincida con la columna real y rechace decimales silenciosamente
  truncados.
- Frontend: quitar la key camelCase radioPermitido residual del body
  enviado (antes viajaba sin uso via el spread de payload).
- Doc: registrar en el spec que la regla implementada difiere de la
  originalmente propuesta (sometimes en vez de nullable, ver razon).

Sin hallazgos Critical/Important en la revision final de todo el batch;
esto cierra los 3 Minor que si eran baratos de resolver.

+++ backend/app/Http/Controllers/Api/TiendaController.php
@@ -45,7 +45,7 @@ class TiendaController extends Controller
             'activo'    => ['boolean'],
             'latitud'   => ['sometimes', 'nullable', 'numeric', 'between:-90,90'],
             'longitud'  => ['sometimes', 'nullable', 'numeric', 'between:-180,180'],
-            'radio_permitido' => ['sometimes', 'numeric', 'min:1'],
+            'radio_permitido' => ['sometimes', 'integer', 'min:1'],
         ]);
 
         $data['activo'] = $data['activo'] ?? true;
@@ -65,7 +65,7 @@ class TiendaController extends Controller
             'activo'    => ['boolean'],
             'latitud'   => ['sometimes', 'nullable', 'numeric', 'between:-90,90'],
             'longitud'  => ['sometimes', 'nullable', 'numeric', 'between:-180,180'],
-            'radio_permitido' => ['sometimes', 'numeric', 'min:1'],
+            'radio_permitido' => ['sometimes', 'integer', 'min:1'],
         ]);
 
         $tienda->update($data);
